Add function aliases in libmsvcrt.a for arm just like on x86/x86_64
Function aliases aren't supported in genlib though, only in
GNU binutils dlltool (which doesn't support arm) and llvm-dlltool.
It didn't make sense to do this until now, when we have a dlltool
that supports arm.

This adds the function aliases from msvcrt-common.def.in, functions
without a leading underscore (such as open, read, close, which
are remapped to the msvcrt functions _open, _read, _close).

Signed-off-by: Martin Storsjö <martin@martin.st>