math: Call the normal double copysign in copysignl on arm
We already have a C based copysign function which is used on arm.
On arm, long double is equal to double, so just call the normal
double function instead of using the handwritten assembly.

The C based copysign function ends up at 5 instructions including
"bx lr", which is significantly shorter than this version.

Signed-off-by: Martin Storsjö <martin@martin.st>