headers: Avoid using the dllimported versions of Interlocked functions on i386
Complete the mapping of unprefixed Interlocked functions to the
ones starting with an underscore in winnt.h.

This makes it consistent with x86_64.

Avoid declaring the dllimported versions of the functions. We
currently export them with the __imp_ prefix in lib32/kernel32.def
(but marked as DATA, so there is no symbol without the __imp_ prefix).
When these functions are used normally, this works just fine, but
if e.g. initializing a global variable pointing to such a function,
it will end up referencing the actual function name without the __imp_
prefix (which lld will fail to link).

This fixes linking sqlite3 as a part of Qt for i386.

Signed-off-by: Martin Storsjö <martin@martin.st>