crt: Fix native TLS with LLVM/LLD
Place the start pointer in ".tls" instead of ".tls$AAA"; LLVM places
tls data in sections named ".tls$" so make sure the start pointer gets
sorted before this.

Don't place the _tls_used directory in the tls segment. This data itself
isn't part of the thread local data but it is only the data descriptor,
which can go in a normal data segment.

This makes sure that the offsets of thread local data symbols as calculated
by the linker with SECREL relocations from the start of the .tls section
match the range starting from _tls_start allocated by the OS and initialized
with the data in the .tls section.

Signed-off-by: Martin Storsjö <martin@martin.st>