crt: Add a def file for vcruntime140_app.dll
This file isn't strictly a part of windows as such, but is a
redistributable runtime for MSVC 2015/2017.

libucrt.a provides functions such as memcmp, and memcmp and memmove as
part of api-ms-win-crt-private-l1-1-0.dll, but it is not allowed to link
to that one from Windows Store apps.

It is allowed to link to vcruntime140_app.dll, though, (since that DLL
is handled as a packaged dependency for the app, not a system API).

Signed-off-by: Martin Storsjö <martin@martin.st>