winstorecompat: Add an implementation of EnumProcessModules
This only works for the local process. This uses structs from
winternl.h that, according to documentation, could change between
Windows versions.

While it legitimately could change, it hasn't so far (this layout
works with XP to 10), and for the Windows Store subset, this seems
like the most pragmatic solution.

llvm's libunwind uses this function for resolving dwarf exception
sections at runtime.

Signed-off-by: Martin Storsjö <martin@martin.st>