Add missed functions to retrieve pseudo-handles
This patch adds functions to retrieve pseudo-handles of process or/and
threads used to specify their tokens.

The functions are missed in the mingw-w64's headers but exist in headers
of Microsoft Windows SDK v10.0.14393.0 and later. According to the
original (Windows SDK's) file ``processthreadsapi.h'' the function are
available starting from Microsoft Windows 8, but I did not found them in
Microsoft Windows SDK 8 header file(s).

Signed-off-by: Ruslan Garipov <ruslanngaripov@gmail.com>
Signed-off-by: Liu Hao <lh_mouse@126.com>