headers: crt: allow blocking some APIs in winstore builds
Based on this documentation some APIs are not allowed in UWP builds
https://docs.microsoft.com/en-us/cpp/cppcx/crt-functions-not-supported-in-universal-windows-platform-apps?view=vs-2019

The documentation is old and may not apply to Win10 but it's still the one
listed for UCRT compatibility :
https://docs.microsoft.com/en-us/cpp/c-runtime-library/compatibility?view=vs-2019

Define _CRT_USE_WINAPI_FAMILY_DESKTOP_APP to let the compiler know the API's are
allowed. So we can restrict some C runtime APIs based on the target family.

Signed-off-by: Liu Hao <lh_mouse@126.com>