winstorecompat: Forward VirtualProtect to VirtualProtectFromApp on windows 10
VirtualProtectFromApp is a fully functional replacement for cases when
not needing to touch code sections, and for other cases, the call
should fail - just like the previous stub that failed unconditionally.

This should be enough for fixing up runtime pseudo relocations for
code generated by LLVM (and possibly by GCC on x86_64), where all
fixed up references either are in normal data sections, or in .refptr
stubs.

Signed-off-by: Martin Storsjö <martin@martin.st>