crt: Check pseudo relocations for overflows and error out clearly
Normally, when correctly configured, the pseudo relocations should
be in fields that are large enough to hold the full target
offset/address. But if the relocations nevertheless end up truncated,
error out clearly instead of running into a hard to diagnose crash
at runtime.

The pseudo relocations can be applied both on absolute pointers and
relative offsets, so when writing a N bit number, we don't know if
the limits for it are unsigned or signed. Thus carefully allow
values from -(2^(N-1)) to (2^N)-1, covering the full range for both
signed and unsigned N bit numbers. This won't catch all cases where
offsets are out of bounds, but should catch the vast majority, allowing
a clearer error message in those situations.

By default, GCC builds for x86_64 with the medium code model, which
adds .refptr stubs when referencing addresses that might end up
autoimported (i.e. when referencing addresses that can be out of range
for a 32 bit offset). Some users, who don't expect to be autoimporting
any data symbols, might be building with -mcmodel=small [1], which
avoids this extra indirection - but which then silently breaks things
if actually ending up autoimporting data symbols from another DLL.

This can also happen if calling a function which is marked "DATA" in
the def files as it's not meant to be called/used normally (because we
provide a replacement in libmingwex or lib*crt* that we think should
be used instead). If the function that is meant to be called is missing
(this can happen in misconfigured builds where the libraries are lacking
symbols that we expect to provide, see [2]), the linker can end up doing
an autoimport of the function into a 32 bit RIP-relative offset.
(This only happens with Clang; GCC creates a .refptr stub for the
function in these cases, while Clang expects such stubs not to be needed
for functions, only for data.)

[1] https://code.videolan.org/videolan/dav1d/-/commit/8f7af99687533d15a9b5d16abc7b9d7b0cd4dcd0
[2] https://github.com/ziglang/zig/issues/9845

Signed-off-by: Martin Storsjö <martin@martin.st>