crt: Provide opengl32 and glu32 import libraries for arm32 and arm64
These DLLs are available out of the box on arm32 and arm64 since
Windows 11.

However these DLLs themselves are only the main interface (containing
the fallback, essentially unusable, OpenGL 1.1 GDI implementation);
without a separate ICD (installable client driver), it's near useless.

Currently (Dec 2021), such a driver (implementing OpenGL on top of
D3D12) is installable in the "OpenCL and OpenGL Compatibility Pack" in
Microsoft Store (but it's only available for regular third party
applications if you install the Windows Insider version of it).
And even then, it's not necessarily available for all architectures.
(In a brief test, it seems like the driver is available for i686 and
aarch64 processes, but not for x86_64 or armv7).

By providing these import libraries, we risk that people porting
software to Windows 10 on ARM accidentally end up making binaries that
use OpenGL, which don't load on Windows 10, only on Windows 11.
(Earlier, they got clear link errors showing that these libraries
really aren't available for these architectures.)

For some projects with existing support for Windows 10 on ARM, the
fact that OpenGL is unavailable is handled by entirely disabling
building of intermediate wrapper libraries such as glew or libepoxy;
then the main application only detects the presence of glew or
libepoxy, and disables OpenGL functionality if not found. (This has
been tested with VLC, where the glew dependency already is hardcoded
to not be built for these targets, so this doesn't change anything for
the build output of VLC.)

As an additional measure, one could consider adding version guards
in the OpenGL headers, making the declarations unavailable when
targeting an older version of Windows that don't have the functions.
That would allow applications to detect whether it's available or
not depending on what Windows version they target. However, that
doesn't seem to be necessary right now.

Signed-off-by: Martin Storsjö <martin@martin.st>