crt: Remove a duplicate build rule for libntoskrnl.a
For the 32 bit x86 target, there's a more specific build rule
for libntoskrnl.a in the toplevel Makefile.am, where it is built
including libsrc/memcmp.c into the end library too.

I'm unsure if this is a mistake (that the x86_64 and arm versions
of this library doesn't include that extra object file) or if
it's meant to be this way. But in any case, remove the currently
duplicate build rule for this file.

This hopefully fixes spurious install failures (where two commands
try to install the same file at the same time).

Signed-off-by: Martin Storsjö <martin@martin.st>