math: Improve precision of pow(x, y) with integer y < 0
Instead of calculating the reciprocal of x upfront and doing the
repeated multiplication with that, calculate the reciprocal of
the result at the end instead.

As a special case, for cases where this algorithm overflows
(where the result would be a subnormal), fall back on the previous
algorithm.

Instead of spending upfront effort on trying to decide between
these two algorithms, just try to detect whether this happened
afterwards (which should be quite cheap) and redo the calculation
if that happened - with the assumption that this is the uncommon
case (subnormals generally can be considered more expensive).

This fixes https://github.com/msys2/MINGW-packages/issues/11733.

Signed-off-by: Martin Storsjö <martin@martin.st>