crt: Modernize the NAN constants returned by the nan*() functions
Don't return NAN constants with the sign bit set.

This matches what UCRT returns for the nan*() functions since
UCRT was introduced. It also matches the _Nan constant provided
in older versions of MSVC (at least since 2003) (which are returned
by various functions).

Only in MSVC 6, I found NAN constants that had the sign bit set,
matching what we had (with the intent to match MSVC).

Thus, I think it's better for us to match the behaviour of modern
MSVC (2003 and forward) than the legacy behaviour of MSVC 6 in
this aspect.

Signed-off-by: Martin Storsjö <martin@martin.st>