crt: Move the strtof() function from libmingwex.a to libmsvcr*.a
This avoids using this function (where strtof() is implemented
with strtof()) in UCRT builds.

In stdlib.h, we don't define __USE_MINGW_STRTOX when targeting
UCRT - with the intent that we'd use the CRT's implementation
instead. This makes that actually work as intended for strtof.

This also fixes accuracy issues reported in
https://github.com/msys2/MINGW-packages/issues/12472 by simply
deferring to the UCRT's implementation - as originally intended.

Signed-off-by: Martin Storsjö <martin@martin.st>