crt: expose non-dllimport variant of rand_s
The rand_s function is declared as dllimport in the headers, so
in general, it should be enough to provide the __imp_rand_s
symbol.

However, LLD's LTO fails if both the object referencing the
dllimported rand_s and the object file providing the implementation
are LTO object files; see
https://github.com/llvm/llvm-project/issues/57982.

(For most traditional mingw-w64 toolchain distributions, the object
files from the mingw-w64 crt are provided as regular object files,
so this issue isn't triggered. However, Zig builds these object
files from scratch individually, and when doing an LTO build, it
builds all of them with LTO enabled too, hitting this issue.)

Signed-off-by: mlugg <mlugg@mlugg.co.uk>
Signed-off-by: Martin Storsjö <martin@martin.st>