crt: Provide the stack protector functions/variables in libmingwex too
While the compiler drivers do add -lssp automatically when
linking with -fstack-protector-strong, this also allows relying
entirely on libmingwex. This implementation should be functionally
equivalent to GCC's libssp for a mingw build configuration.

This allows toolchains that don't otherwise use GCC's runtime libraries
(such as LLVM based ones) to skip building libssp from GCC (and
replacing it with empty/dummy libssp.a and libssp_nonshared.a).

Signed-off-by: Martin Storsjö <martin@martin.st>