headers/_mingw: Remove the obsolete `__stdcall` macro
This was introduced in 8bcbcee35c14048ab361c1f4db12dbd5b521cd07.

Both GCC and Clang have it as a builtin macro, which expands to
`__attribute__((__stdcall__))`. It shouldn't be redefined, unless it's
really necessary, for example, when working around bugs in a certain
version of the compiler.

Nowadays, in the case of x86_64, this code

  int (* __cdecl cdecl_ptr)(int);
  int (* __stdcall stdcall_ptr)(int) = cdecl_ptr;

compiles without warnings or errors. The `__stdcall` attribute does
not have any effects whatsoever and can be left intact.

Signed-off-by: LIU Hao <lh_mouse@126.com>