crt: Handle TLS destructors that are registered while executing destructors
Instead of storing a pointer to a dtor_obj in the TLS key, store
a pointer to a pointer as a stable head of the list. This allows
the TLS key to be set only once, keeping just the header pointer
stable and updating the header pointer to point to the latest
addition.

When running destructors, update the header pointer continuously
(before running each destructor), to allow it to be updated during
the call. (I.e., run_dtor_list is rewritten to use *ptr in each
step instead of iterating over a local variable.)

This fixes a libcxxabi test case at
https://github.com/llvm/llvm-project/blob/llvmorg-16.0.0/libcxxabi/test/thread_local_destruction_order.pass.cpp.

Signed-off-by: Martin Storsjö <martin@martin.st>