headers: Define IMAGE_RUNTIME_FUNCTION_ENTRY for aarch64
We've had a definition of IMAGE_ARM64_RUNTIME_FUNCTION_ENTRY since
a1fc845070d144bd0052cfccbe91195646eb3bec in 2018, but we missed
the definition of the generic IMAGE_RUNTIME_FUNCTION_ENTRY to
point to the arch specific version of it.

The shorter names RUNTIME_FUNCTION and PRUNTIME_FUNCTION already
are typedefs for IMAGE_ARM64_RUNTIME_FUNCTION_ENTRY though.

Signed-off-by: Martin Storsjö <martin@martin.st>